added text to each tile, that tells how many bombs are around
