+++ main.cpp
@@ -23,7 +23,7 @@ int main()
     TileMap tile_map(lvlSize, tileTexture);
     tile_map.load("tile.png", sf::Vector2u(64, 64), tile_map.getTiles(), lvlSize, lvlSize, sf::Vector2f(0.f, 100.f));
     tile_map.createBombs(10);
-    tile_map.uncoverWholeMap();
+    //tile_map.uncoverWholeMap();
 
     //sf::Font font;
     //if (!font.loadFromFile("ArialTh.ttf"))
@@ -34,6 +34,7 @@ int main()
     //sf::Text text;
     //text.setFont(font);
     //text.setString("Hello world");
+    //text.setPosition(sf::Vector2f(0.f, 0.f));
     //text.setCharacterSize(24);
     //text.setFillColor(sf::Color::Red);
     //text.setStyle(sf::Text::Bold | sf::Text::Underlined);
@@ -69,7 +70,7 @@ int main()
                 window.draw(tile_map.tiles[i][j].bombsAroundText);
             }
         }*/
-        //tile_map.drawText(window);
+        tile_map.drawText(window);
         window.display();
     }
